+++ src/Model/Table/ServiceTypesTable.php
@@ -78,11 +78,11 @@ class ServiceTypesTable extends AppTable
             ->allowEmptyString('subscriber_verification_code_format');
 
         $validator
-            ->integer('activation_fee')
+            ->decimal('activation_fee')
             ->allowEmptyString('activation_fee');
 
         $validator
-            ->integer('activation_fee_with_obligation')
+            ->decimal('activation_fee_with_obligation')
             ->allowEmptyString('activation_fee_with_obligation');
 
         $validator
